future_transactions/requirements.txt
flask
flask_api
requests==2.27.1
pandas==1.1.5
PyYAML==6.0.1